refactor(a11y): extract the drawer focus trap into a shared hook
Lifts the editor drawer's focus trap verbatim, container-focused
Shift+Tab edge case included, so the new global nav reuses it rather
than hand-rolling a weaker one. Behaviour unchanged.
